+++ mqtt-sn.c
@@ -353,7 +353,7 @@ static uint8_t mqtt_sn_get_qos_flag(int8_t qos)
     }
 }
 #if 1
-void mqtt_sn_send_publish(struct mqtt_sn_connection *mqc, uint16_t topic_id, uint8_t topic_type, const char* data, size_t data_len, int8_t qos, uint8_t retain)
+void mqtt_sn_send_publish(struct mqtt_sn_connection *mqc, uint16_t topic_id, uint8_t topic_type, const char* data, uint16_t data_len, int8_t qos, uint8_t retain)
 {
     publish_packet_t packet;
